feat: earthly

+++ src/webdrivers/common/tar_gz_extractor.cr
@@ -9,10 +9,12 @@ class Webdrivers::Common::TarGzExtractor
   def extract
     Compress::Gzip::Reader.open(file) do |gzip|
       Crystar::Reader.open(gzip) do |tar|
-        entry = tar.next_entry.not_nil!
-        destination_path = File.join(install_path, entry.name)
-        File.delete(destination_path) if File.exists?(destination_path)
-        File.write(destination_path, entry.io)
+        if next_entry = tar.next_entry
+          entry = next_entry
+          destination_path = File.join(install_path, entry.name)
+          File.delete(destination_path) if File.exists?(destination_path)
+          File.write(destination_path, entry.io)
+        end
       end
     end
   end
